requirements.txt
asgiref
beautifulsoup4~=4.9.1
Django==2.2.14
google~=2.0.3
lxml
py
pytz
soupsieve
sqlparse
Pillow
django-bootstrap4-form
celery
django-celery-beat
email-reply-parser
akismet
markdown
google-api-python-client
simplejson
six~=1.15.0
djangorestframework
django-model-utils~=4.0.0
pinax-templates
uvloop~=0.14.0
websockets~=8.1
django-braces~=1.14.0
django-jet
FeedParser
bootstrap4
django-jsonfield~=1.4.0
requests~=2.24.0
django-mssql-backend
django-ipware
psycopg2
django-mssql
